+++ extcss3/intern.h
@@ -26,9 +26,9 @@ void extcss3_release_block(extcss3_block *block);
 void extcss3_release_decl(extcss3_decl *decl);
 void extcss3_release_decls_list(extcss3_decl *list);
 
-bool extcss3_set_css_string(extcss3_intern *intern, char *css, size_t len, int *error);
-bool extcss3_set_modifier(extcss3_intern *intern, short int type, void *callable, int *error);
-bool extcss3_set_vendor_string(extcss3_intern *intern, char *name, size_t len, int *error);
+bool extcss3_set_css_string(extcss3_intern *intern, char *css, size_t len, unsigned int *error);
+bool extcss3_set_modifier(extcss3_intern *intern, unsigned int type, void *callable, unsigned int *error);
+bool extcss3_set_vendor_string(extcss3_intern *intern, char *name, size_t len, unsigned int *error);
 
 /* ==================================================================================================== */
