+++ extcss3/minifier/types/function.c
@@ -9,12 +9,12 @@
 
 /* ==================================================================================================== */
 
-bool extcss3_minify_function_rgb_a(extcss3_token **token, extcss3_decl *decl, int *error)
+bool extcss3_minify_function_rgb_a(extcss3_token **token, extcss3_decl *decl, unsigned int *error)
 {
 	extcss3_token *temp, *curr = (*token)->next;
 	double value;
 	char hex[9];
-	unsigned short int idx = 0, percentages = 0, numbers = 0;
+	unsigned int idx = 0, percentages = 0, numbers = 0;
 	bool valid = true;
 
 	/* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
